[FMA-117] 게시글 수정 오류

+++ app/src/main/java/io/familymoments/app/feature/addpost/uistate/AddPostUiState.kt
@@ -14,7 +14,7 @@ data class AddPostUiState(
 
 @Immutable
 data class ExistPostUiState(
-    val editPostId: Int = 0,
+    val editPostId: Long = 0,
     val editImages: List<String> = listOf(),
     val editContent: String = ""
 )
